SqZpMatrix

Description

is a matrix of ZpInt's with the main purpose of this module being Gaussian elimination and kernel calculation. Here, p must be prime.

Documentation

newtype SqZpMatrix

Constructors

SqZpMatrix [[ZpInt]]

getMatrixData :: SqZpMatrix -> [[ZpInt]]

buildMatrix :: [[Int]] -> Int -> SqZpMatrix

Constructs a SqZpMatrix given a 2D array of ints and desired modulus

rref :: SqZpMatrix -> (SqZpMatrix, [Int])

Perform Gaussian elimination until matrix is in reduced echelon form with indexes for pivot columns

dimensions :: SqZpMatrix -> (Int, Int)

Returns (num_rows, num_cols)

sqTranspose :: SqZpMatrix -> SqZpMatrix

Transposes a matrix

nullspace :: SqZpMatrix -> [[ZpInt]]

Calculates the kernel, or nullspace, of a ZpInt Matrix