Utility

Description

module holding various useful general functions

Documentation

xor :: Bool -> Bool -> Bool infix 2

Exclusive OR statement

countSuchThat :: (a -> Bool) -> [a] -> Int

Count elements in [a] that return True for (a -> Bool)

doKTimes :: (a -> a) -> Int -> a -> a

Compose a function with itself a given amount of times

slideWindow :: [a] -> Int -> [[a]]

Slides a selection window along a given list, with given window size

enumerate :: [a] -> [(Int, a)]

Enumerates a list and send [a, b, c, ...] to [(1, a), (2, b), (3, c), ...]

getRange :: [a] -> Int -> Int -> [a]

Gets the elements of a list whose indexes are in the interval [i1, i2]