ZpCEdga

Description

Interface for handling Chekanov-Eliashberg differential graded algebras directly. Only gradings allowed are Z_p[t^(-1), t] for prime p

Documentation

type Augmentation = [ZpInt]

newtype Differential

Constructors

Differential

Fields

getDiffPairs :: [(Expression, Expression)]

data DGA

Constructors

DGA

Fields

gradings :: [ZpInt]

differentials :: Differential

getFroms :: Differential -> [Expression]

Returns the generators which can be differentiated

getTos :: Differential -> [Expression]

Returns the results of the differential applied the generators

differentialFromStrings :: Int -> [(String, String)] -> Differential

Constructs a Differential type given a string [(to's, from's)] applying expressionFromString to each element element of each pair

exprGen :: [Expression] -> Int -> [Expression]

Takes expressions and returns a list of terms up to given length where any element in returned list is from any term in the list

diff :: DGA -> Expression -> Expression

Can only differentiate single elements. If you want to differentiate in multi-element terms linearly, use diff_linear. Otherwise, use substitute.

diff_linear :: Differential -> Expression -> Expression

Differentiates but only linearly, meaning no chain-rule! This can handle products of terms that aren't explicitly listed in the Differential. The purpose of this function is to double-check that d^2 = 0 in a polynomial DGA

getDGAModulus :: DGA -> Int

Gets the modulus of the first term in the first element of the generator of the DGA

buildDGA :: Int -> [Int] -> [(String, String)] -> DGA

Builds a DGA given gradings, modulus, (tos, froms)

grading :: DGA -> Expression -> ZpInt

Computes the grading of an expression in the DGA. WARNING: Only actually computes the grading of the first term in an expression, since such determines the grading of the whole expression

augmentations :: DGA -> [Augmentation]

Computes all possible graded augmentations for given DGA

polyDGA :: DGA -> Augmentation -> Int -> DGA

Takes a DGA and increases num of generators appropriately, as well as well as restricts boundary map to words of length specified TODO: Function fails on polyDGA (kUnknot_iso1 3) (head $ augmentations $ kUnknot_iso1 3) 2

homology :: DGA -> Int -> Int

Computes the homology of a DGA which has already been passed to polyDGA

testHomology :: DGA -> Int -> Int -> Int

Calculates lth homology of (polyDGA k) with first augmentation found

polyHomologies :: DGA -> Int -> [[Int]]

Computes all gradings of homology of all possible augmentations of polyDGA given degree

kUnknot :: Int -> DGA

kUnknot_iso1 :: Int -> DGA

kTrefoil :: Int -> DGA

km10_140 :: Int -> DGA

kK2 :: Int -> DGA