ZpExpression

Description

containing all of the relevant algebraic manipulations to expressions, such as addition, division, substitution, etc. ZpExpression requires coefficients to be of the ZpInt type.

Documentation

type Term = [String]

data Expression

Constructors

Expression [ZpInt] [Term]

coefficients :: Expression -> [ZpInt]

Gets coefficients of expression

terms :: Expression -> [Term]

Gets terms of expression

getCoeffTermPairs :: Expression -> [(ZpInt, Term)]

Gets list of pairs of (coeff, term) from expression

numTerms :: Expression -> Int

Calculates the number of terms in an expression

getTermCoeffInExpression :: Term -> Expression -> ZpInt

Gets the coefficient of a given term in an expression. Returns -1 mod -1 if no coefficient is found

getTermIndex :: Term -> Expression -> Int

Gets the index of a given term in an expression

termInExpr :: Term -> Expression -> Bool

Tells whether or not a certain term is in an expression. WARNING: This checks for the entire term. For example, termInExpr "q1" "q1*q2" = FALSE but termInExpr "q1" "q1+q2" = TRUE

getExprModulus :: Expression -> Int

Gets modulus of the coefficient of the first term

headExpr :: Expression -> Expression

Gets the first term in an Expression and returns it as its own expression

tailExpr :: Expression -> Expression

Returns the expression minus the first term

expressionFromString :: String -> Int -> Expression

Converts a string of the form "q1*q2 + 2*q3*3" into (Expression [1, 6] [["q1", "q2"], ["q3"]]) and reduced given modulus

substitute :: [Term] -> [Expression] -> Expression -> Expression

Substitutes corresponding terms with given expressions (a homomorphism-type map with [Term] -> [Expression]). First attempts to substitute entire terms. If such fails, substitute single elements of each term one by one.

restrictTermLen :: Expression -> Int -> Expression

Restricts expression to terms of given length